% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/govspeak.R
\name{application_js_dependency}
\alias{application_js_dependency}
\title{Function to add application.js as a dependency}
\usage{
application_js_dependency()
}
\value{
A HTML dependency object for application.js
}
\description{
This function adds application.js as a dependency in an HTML document.
}
