% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/govspeak.R
\name{check_for_figure_html_dir}
\alias{check_for_figure_html_dir}
\title{Check for figure-html directory in auxiliary files}
\usage{
check_for_figure_html_dir(path)
}
\arguments{
\item{path}{The path to check for auxiliary files.}
}
\value{
TRUE if a figure-html directory is found, FALSE otherwise.
}
\description{
Check for figure-html directory in auxiliary files
}
