% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/govspeak.R
\name{clean_up}
\alias{clean_up}
\title{Clean up environment state after rendering}
\usage{
clean_up(input_file)

clean_up(input_file)
}
\arguments{
\item{input_file}{The input file to be processed (currently unused).}
}
\value{
None. Resets env_state variables.

No return value; called for side effects.
}
\description{
Resets the `file_names` and `format` variables in the `env_state` environment.
}
\details{
This function clears the `file_names` vector and sets the `format` variable to `NULL`
within the `env_state` environment. It is typically used to reset the state before
processing a new input file.
}
\keyword{internal}
