% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/govspeak.R
\name{convert_callouts}
\alias{convert_callouts}
\title{Convert markdown callout syntax to govspeak}
\usage{
convert_callouts(md_file)
}
\arguments{
\item{md_file}{string; markdown file text}
}
\description{
Convert markdown callout sytax to govspeak
}
