% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/govspeak.R
\name{convert_image_references}
\alias{convert_image_references}
\title{Convert markdown image references to govdown}
\usage{
convert_image_references(image_references, md_file, images_folder)
}
\arguments{
\item{image_references}{dataframe of image file names and associated govdown
reference.}

\item{md_file}{string with markdown file text}

\item{images_folder}{string; folder containing images for *.md file}
}
\description{
Convert markdown image references to govspeak format (!!n)
}
