% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/govspeak.R
\name{convert_md}
\alias{convert_md}
\title{Convert Markdown to a govspeak file}
\usage{
convert_md(input_file, renamed_images)
}
\arguments{
\item{input_file}{The path to the input Markdown file.}

\item{renamed_images}{Data frame of renamed images (optional).}
}
\value{
The govspeak file as a string.
}
\description{
Convert Markdown to a govspeak file
}
