% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/govspeak.R
\name{generate_image_references}
\alias{generate_image_references}
\title{Generate govdown image references}
\usage{
generate_image_references(img_filenames)
}
\arguments{
\item{img_filenames}{character vector of files to be referenced in govdown format
(!!n). The filename must start and end with a number and have text in
between (eg, "1-abcd-1.png")}
}
\description{
Converts image file names to a dataframe, with a field containing the
original image name and corresponding govdown reference
}
