% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/govspeak.R
\name{govspeak}
\alias{govspeak}
\title{Render a document using the GOV.UK rgovspeak format}
\usage{
govspeak(
  image_type = "svg",
  fig_width = 960L/72L,
  fig_height = 640L/72L,
  dpi = 72,
  pandoc_args = NULL,
  keep_md = FALSE,
  ...
)
}
\arguments{
\item{image_type}{The image type for figures (default: "svg").}

\item{fig_width}{Figure width in inches (default: 960/72).}

\item{fig_height}{Figure height in inches (default: 640/72).}

\item{dpi}{Figure resolution in dots per inch (default: 72).}

\item{pandoc_args}{Additional Pandoc arguments.}

\item{keep_md}{Logical; whether to keep the intermediate markdown file.}

\item{...}{Additional arguments passed to rmarkdown::html_document_base.}
}
\value{
An rmarkdown_output_format object.
}
\description{
Render a document using the GOV.UK rgovspeak format
}
