% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/govspeak.R
\name{govspeak_dependency}
\alias{govspeak_dependency}
\title{Function to create a HTML dependency for Govspeak templates}
\usage{
govspeak_dependency()
}
\value{
A HTML dependency object for Govspeak templates
}
\description{
This function creates a HTML dependency for Govspeak templates, which includes the name, version,
source, and stylesheet.
}
