% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/govspeak.R
\name{move_image_files_to_extension_dirs}
\alias{move_image_files_to_extension_dirs}
\title{Move image files to extension-based subdirectories under ../../images}
\usage{
move_image_files_to_extension_dirs(image_dir)
}
\arguments{
\item{image_dir}{The directory containing image files to move.}
}
\value{
None. Files are moved on disk.
}
\description{
Move image files to extension-based subdirectories under ../../images
}
