% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/govspeak.R
\name{pkg_file}
\alias{pkg_file}
\title{Function to get the file path of a resource file in the package}
\usage{
pkg_file(...)
}
\arguments{
\item{...}{additional path components to the resource file}
}
\value{
the file path of the resource file
}
\description{
Function to get the file path of a resource file in the package
}
