% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/govspeak.R
\name{remove_header}
\alias{remove_header}
\title{Replace R markdown header with ## title}
\usage{
remove_header(md_file)
}
\arguments{
\item{md_file}{string; markdown file text}
}
\description{
Replace R markdown header with title only
}
