% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/govspeak.R
\name{remove_rmd_blocks}
\alias{remove_rmd_blocks}
\title{Remove R markdown multiline block elements (package warning and code block)}
\usage{
remove_rmd_blocks(md_file)
}
\arguments{
\item{md_file}{string; markdown file text}
}
\description{
Remove R markdown multiline block elements (package warnings, but also multiline code blocks)
}
