% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/govspeak.R
\name{rename_images}
\alias{rename_images}
\title{Rename images in a directory by appending a date string}
\usage{
rename_images(image_dir, date_str)
}
\arguments{
\item{image_dir}{The directory containing image files.}

\item{date_str}{The date string to append to filenames.}
}
\value{
A data frame mapping original to new filenames.
}
\description{
Rename images in a directory by appending a date string
}
