% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/govspeak.R
\name{save_data}
\alias{save_data}
\title{Save data and track file names for post-processing}
\usage{
save_data(save_func, args_list)
}
\arguments{
\item{save_func}{The function to use for saving data.}

\item{args_list}{List of arguments to pass to save_func.}
}
\value{
The result of save_func.
}
\description{
Save data and track file names for post-processing
}
