% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/govspeak.R
\name{govspeak}
\alias{govspeak}
\title{FUNCTION_TITLE}
\usage{
govspeak(
  fig_width = 960/72,
  fig_height = 640/72,
  fig_caption = TRUE,
  dpi = 72,
  toc = TRUE,
  ...
)
}
\arguments{
\item{fig_width}{PARAM_DESCRIPTION, Default: 960/72}

\item{fig_height}{PARAM_DESCRIPTION, Default: 640/72}

\item{fig_caption}{PARAM_DESCRIPTION, Default: TRUE}

\item{dpi}{PARAM_DESCRIPTION, Default: 72}

\item{toc}{PARAM_DESCRIPTION, Default: TRUE}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
